requirements.txt
aiohttp==3.7.4.post0
alibabacloud-credentials==0.1.4
alibabacloud-dysmsapi20170525==2.0.2
alibabacloud-endpoint-util==0.0.3
alibabacloud-openapi-util==0.1.4
alibabacloud-tea==0.2.7
alibabacloud-tea-openapi==0.2.5
alibabacloud-tea-util==0.3.4
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.8.2
billiard==3.6.4.0
celery==5.1.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cobble==0.1.3
colorama==0.4.4
configparser==5.0.2
construct==2.5.3
crayons==0.4.0
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.20
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-haystack==3.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-mdeditor==0.1.18
django-pandas==0.6.4
django-simpleui==2021.4.1
djangorestframework==3.9.1
drf-haystack==1.8.10
et-xmlfile==1.0.1
future==0.18.2
idna==2.10
jieba==0.42.1
kombu==5.1.0
loguru==0.4.1
lxml==4.6.3
mammoth==1.4.13
Markdown==3.3.3
markdownify==0.6.0
MarkupPy==1.14
multidict==5.1.0
numpy==1.20.2
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.2.4
pefile==2019.4.18
Pillow==7.1.0
prompt-toolkit==3.0.18
pycparser==2.20
pydantic==1.8.1
pyee==8.1.0
PyMySQL==1.0.2
pyperclip==1.8.2
pyppeteer==0.0.25
python-dateutil==2.8.1
python-ptrace==0.9.8
pytz==2021.1
PyYAML==5.4.1
requests==2.24.0
rope==0.19.0
selenium==3.141.0
setuptools-scm==6.0.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.3.0
websockets==8.1
Whoosh==2.7.4
xlrd==2.0.1
xlwt==1.3.0
yapf==0.31.0
yarl==1.6.3
